refactor: 파일, 폴더 조금 수정
